restructured project, improved concruency
