+++ abstract_factory/abstract_factory.py
