--- entities/list_objects.py
@@ -1,13 +1,0 @@
-class ConcessionSales:
-    def __init__(self, concession_sales_info):
-        self.concession_sales_info = concession_sales_info
-
-    def get_concession_sales_info(self):
-        return self.concession_sales_info
-
-class TicketSales:
-    def __init__(self, ticket_sales_info):
-        self.ticket_sales_info = ticket_sales_info
-
-    def get_ticket_sales_info(self):
-        return self.ticket_sales_info
