+++ logic_pkg/concession_sales_logic.py
@@ -1,7 +1,8 @@
 from os import read
 import PySimpleGUI as sg
 from controllers import ui_controller, logic_controller
-from entities import listings, facade
+from bridge import listing_factory
+from facade import facade
 
 reader = facade.Reader()
 
@@ -12,7 +13,7 @@ def concessionSalesEventLoop(window, event, values):
         sg.popup('{} sales'.format(values['-List-'][0]))
     if event == 'Update Values':
         concession_sales_info = reader.read("databases/concession_sales_db.txt", "")
-        concession_sales_list = listings.list_factory.create_list("concession sale", concession_sales_info)
+        concession_sales_list = listing_factory.factory.create_list("concession sale", concession_sales_info)
         concession_sales_screen = concession_sales_list.generate_list()
         window['-List-'].update(values=concession_sales_screen)
