+++ logic_pkg/ticket_sales_logic.py
@@ -1,7 +1,7 @@
 import PySimpleGUI as sg
 from controllers import ui_controller, logic_controller
 from utils import utils
-from entities import listings
+from bridge import listing_factory
 
 def ticketSalesEventLoop(window, event, values):
     if event == 'Back To Menu':
@@ -10,7 +10,7 @@ def ticketSalesEventLoop(window, event, values):
         sg.popup('{} tickets'.format(values['-List-'][0]))
     if event == 'Update Values':
         ticket_sales_info = utils.get_view_list("ticket sale","databases/ticket_sales_db.txt")
-        ticket_sales_list = listings.list_factory.create_list("ticket sale", ticket_sales_info)
+        ticket_sales_list = listing_factory.factory.create_list("ticket sale", ticket_sales_info)
         ticket_sales_screen = ticket_sales_list.generate_list()
         window['-List-'].update(values=ticket_sales_screen)
