+++ ui_pkg/concession_sales_ui.py
@@ -1,11 +1,10 @@
 import PySimpleGUI as sg
-from entities import listings, facade
+from bridge import listing_factory
+from facade import facade
 
 Heading = "TheAter Concession Sales"
 reader = facade.Reader()
-concession_sales_info = reader.read("databases/concession_sales_db.txt", "")
-
-concession_sales_list = listings.list_factory.create_list("concession sale", concession_sales_info)
+concession_sales_list = reader.read("databases/concession_sales_db.txt", "")
 concession_sales_screen = concession_sales_list.generate_list()
 
 adminLayout = [[sg.Text("TheAter Concession Sales")],
