+++ utils/utils.py
@@ -1,108 +1,4 @@
-from entities import list_objects
-from entities import abstract_factory
-
-class ConcessionReader:
-    
-    def read(self, filename):
-        data = open(filename, "r")
-        list = []
-        for line in data: 
-            currentline = line.split(",")
-            # Remove empty items
-            currentline = currentline[:-1]
-            list.append(currentline)
-        return list
-
-class ConcessionSaleReader:
-
-    def read(self, filename):
-            listData = open(filename, "r")
-            lineData = listData
-            list_data = []
-            entryListing = 0
-            for line in lineData:
-                index = 0
-                length = len(line)
-                current_entry = ["","",[]]
-                while index < length:
-                    if line[index] == ",":
-                        index += 1
-                        break
-                    else:
-                        current_entry[0] +=line[index]
-                    index += 1
-                while index < length:
-                    if line[index] == ",":
-                        index += 1
-                        break
-                    else:
-                        current_entry[1] +=line[index]
-                    index += 1
-                entryListing = list_objects.ConcessionSales(current_entry)
-                list_data.append(entryListing)
-            
-            return list_data
-
-class TicketSaleReader:
-
-    def read(self, filename):
-            listData = open(filename, "r")
-            lineData = listData
-            list_data = []
-            entryListing = 0
-            for line in lineData:
-                index = 0
-                length = len(line)
-                current_entry = ["","",[]]
-                while index < length:
-                    if line[index] == ",":
-                        index += 1
-                        break
-                    else:
-                        current_entry[0] +=line[index]
-                    index += 1
-                while index < length:
-                    if line[index] == ",":
-                        index += 1
-                        break
-                    else:
-                        current_entry[1] +=line[index]
-                    index += 1
-                entryListing = list_objects.TicketSales(current_entry)
-                list_data.append(entryListing)
-            
-            return list_data
-
-class BookingsReader:
-
-    def read(self, filename, user):
-        bookings = []
-        with open(filename, "r") as db:
-            for line in db:
-                string = line.split(",")
-                if user == string[0]:
-                    bookings.append(string[1] + ": " + string[2])
-        
-        return bookings
-
-class BookingsReviewReader:
-
-    def read(self, filename):
-        dict = {}
-        with open(filename, "r") as db:
-            for line in db:
-                movie = line.split(",")[1]
-                if movie not in dict:
-                    dict[movie] = 1
-                else:
-                    dict[movie] = dict[movie] + 1
-                
-        dictlist = []
-        for key, value in dict.items():
-            temp = f"{key}: {value}"
-            dictlist.append(temp)
-        
-        return dictlist
+from abstract_factory import abstract_factory
 
 # true for ui format, false for database format
 def title_times_split(string, ui):
